sqlalchemy GET /v1/resources returns too much data
Meters data needs to match mongodb implementation.
The resources returned has meter attribute nested and
those results are each and every row of the meters for
that resource as well as all fields of each row. If a
resource has tens of thousands of meters associated to
it, the eager loading will load all of those rows and
return all that data in the json result. Instead, it
should match the mongodb implementation which simply
returns a distinct list of unique {'counter_name':
value, 'counter_type': value, 'counter_unit': value}.